dev-requirements.txt
nose>=0.11.1
z3c.recipe.tag>=0.4.0
virtualenv>=1.3.3
